// console/src/generated/sdk/ModelDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CatalogProviderDto } from "./CatalogProviderDto";

export type ModelDto = { providers: Array<CatalogProviderDto>, id: string, name: string, metadata: unknown, vocabularyFileId: string | null, };